add CONFIG for unprivileged_userfaultfd
When disabled, unprivileged users will not be able to use the userfaultfd
syscall. Userfaultfd provide attackers with a way to stall a kernel
thread in the middle of memory accesses from userspace by initiating an
access on an unmapped page. To avoid various heap grooming and heap
spraying techniques for exploiting use-after-free flaws this should be
disabled by default.

This setting can be overridden at runtime via the
vm.unprivileged_userfaultfd sysctl.

Signed-off-by: Levente Polyak <levente@leventepolyak.net>